# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Tangula API接口文档\n唐古拉山脉，长江正源，滋润万物。\nTangula 寓意数据的源头，为研发团队提供即取即用的数据服务，帮助团队高效产出。",
        "title": "Tangula",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "zhang.jiejun@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "paths": {
...
        "/log/record": {
            "get": {
                "description": "【登录权限】获取所有操作日志记录，普通用户只能获取自身的操作日志记录，系统管理员平台上所有用户的操作日志记录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "日志管理"
                ],
                "summary": "获取所有日志记录",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "分页索引",
                        "name": "index",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 15,
                        "description": "分页数量",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "用户名称",
                        "name": "user",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/api.Paging"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/log.LogRecord"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/store_pool/image": {
            "get": {
                "description": "【登录权限】获取所有镜像信息，普通用户与系统管理员只能获取公共镜像或自身创建的镜像信息，超级管理员可以获取平台上所有的镜像信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "镜像管理"
                ],
                "summary": "获取所有镜像信息",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "分页索引",
                        "name": "index",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 15,
                        "description": "分页数量",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "镜像权限过滤参数[0.不过滤类型，111.私有镜像，222.公有镜像]",
                        "name": "auth",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "镜像类型过滤参数[0.不过滤类型，1001.FILE，1002.ISO，1003.CUSTOME，1004.VMWAREVM，1005.CASVM，1006.HCSVM，1007.OPENSTACKVM]",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "名称搜索过滤参数",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/api.Paging"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/image.Image"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/store_pool/image/{id}": {
            "get": {
                "description": "【登录权限】获取指定镜像信息，普通用户与系统管理员只能获取公共镜像或自身创建的镜像信息，超级管理员可以获取平台上所有的镜像信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "镜像管理"
                ],
                "summary": "获取指定镜像",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标镜像ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "$ref": "#/definitions/image.Image"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "【登录权限】删除指定镜像，普通用户自能删除属于自身的镜像，管理员可删除自身及公有镜像，超级管理员可删除所有镜像。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "镜像管理"
                ],
                "summary": "删除镜像",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标镜像ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Result"
                        }
                    }
                }
            }
        },
        "/store_pool/image/{id}/publish": {
            "post": {
                "description": "【登录权限】发布指定镜像，只能发布属于用户自身的镜像。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "镜像管理"
                ],
                "summary": "发布镜像",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标镜像ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "$ref": "#/definitions/image.Image"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/store_pool/image/{id}/replica": {
            "post": {
                "description": "【登录权限】通过指定镜像生成副本。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "镜像管理"
                ],
                "summary": "生成副本",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标镜像ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "副本名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "副本描述",
                        "name": "desc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "$ref": "#/definitions/replica.Replica"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Paging": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "totalNum": {
                    "type": "integer"
                }
            }
        },
        "api.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "response": {
                    "type": "object"
                }
            }
        },
...
        "image.Image": {
            "type": "object",
            "properties": {
                "auth_type": {
                    "type": "integer"
                },
                "create_user": {
                    "type": "string"
                },
                "created_time": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pool": {
                    "type": "object",
                    "$ref": "#/definitions/pool.StorePool"
                },
                "pool_id": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "log.LogRecord": {
            "type": "object",
            "properties": {
                "created_time": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "object": {
                    "type": "string"
                },
                "operation": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "user": {
                    "type": "string"
                }
            }
        },
...
        "pool.StorePool": {
            "type": "object",
            "properties": {
                "create_user": {
                    "type": "string"
                },
                "created_time": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "replica.Replica": {
            "type": "object",
            "properties": {
                "create_user": {
                    "type": "string"
                },
                "created_time": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "export": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pool": {
                    "type": "object",
                    "$ref": "#/definitions/pool.StorePool"
                },
                "pool_id": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
    }
}